Conflicts:
	webhooks/gclWebhookHandlers/bill/all.php
	webhooks/gclWebhookHandlers/bill/cancelled.php
	webhooks/gclWebhookHandlers/bill/failed.php
	webhooks/gclWebhookHandlers/bill/paid.php
	webhooks/gclWebhookHandlers/pre_authorization/all.php
	webhooks/gclWebhookHandlers/subscription/all.php
	webhooks/gclWebhookHandlers/subscription/cancelled.php

+++ form_handlers/reset_two_factor.php
@@ -1,4 +1,5 @@
 <?php
+if (!INCLUDED) exit;
 
 session_start();
 delete_user_meta(get_current_user_id(), '2FA_secret');
